Add property-changing JVMTI agent wrapper.
Adds libwrapprops agent that allows one to change or add to the system
properties observed by a JVMTI agent. This is useful for compatibility
as some JVMTI agents expect to be able to find system properties that
are not present in ART. With this we can run them without having to
modify the runtime itself.

Also adds a --agent-wrapper flag to run-jdwp-tests.sh to facilitate
the use of this agent.

Test: mma -j40 libwrapprops
Test: ./art/tools/run-jdwp-tests.sh \
        --mode=host \
        --variant=x64 \
        -Xplugin:$ANDROID_HOST_OUT/lib64/libopenjdkjvmtid.so \
        --jdwp-path /usr/lib/jvm/default-java/jre/lib/amd64/libjdwp.so \
        --agent-wrapper $ANDROID_HOST_OUT/lib64/libwrapprops.so=$PWD/art/tools/libjdwp-compat.props

Change-Id: Ic5eef7e9b4b8c54f4b0683dbb4e71768cbf4f97c