Update patch set 1
Patch Set 1:

> (1 comment)

Please note that the proposal was never to remove GVN, merely to add it to the other two architectures. But I see running the arm-simplifier first is essential.

In that case we still could have a shared path *after* that, so we will have

shared-opts
arch-specific-opts
shared-opts
arch-specific-codegen

to make sure a lot of the code is in common, but perhaps people oppose that idea. Nicolas, Vladimir? What do you think? In the end I would of course also be okay doing it like this cl (and also for x86 then), but I like the idea of having all our compiler paths as similar as possible.

Patch-set: 1
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>