Ensure that OSR doesn't break local-variable get/set
We had a bug where we would on-stack replace a method with a modified
local variable. Thanks to inlining local variable values this could
cause the change to the variable to be lost. We fixed this by giving
plugins a way to notify the runtime they are interested in a particular
method.

Bug: 66959663
Bug: 66933582

Test: while ./test/run-test --host --prebuild -O --jit 1935; do; done
Test: ./test.py --host -j50
Change-Id: Ic001b8a9d8d0bd9ce292e807752c86a505f85d36