Update patch set 2
Patch Set 2:

> I have to integrate dexopt.cpp to make right calls to profman. In a
 > meanwhile, could you please take a look at the current flow?

We should always verify them so let's not pass the --verify flag and we should do it in the assistant.

However, thinking more about this change... we will end up extracting the apk which is not ideal... We could try to get the vdex file but that needs the runtime right now and needs to check the file is up to date.

I expected we can do this in a simpler way and now I wonder if it's worth the complexity...

Patch-set: 2
CC: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>