Fix race in method jitting for 597-deopt-invoke-stub
597-deopt-invoke-stub tests deopting quick-to-interpreter bridge.
It requires a compiled method to call into an uncompiled method.
However the method that's not supposed to be compiled gets compiled
occassionally. It's because that AddSamples() is called twice for the
method, one for Jit::NotifyInterpreterToCompiledCodeTransition()
and one for Jit::MethodEntered(). This triggers the jit
compilation of the method, and when compilation happens very fast,
jit->CanInvokeCompiledCode() picks up the freshly compiled code.

The change bumps up the compilation threshold to 10 in the run script,
instead of the default value of 2 under debug build.

Also exclude speed profile testing configuration for this test. It
generates random profile which may compile the method also.

Test: m -j32 test-art-host-run-test
Bug: 66936353

Change-Id: I119b02a39b8c1b0619fc9fd0ccc51fa034099f7f