Update patch set 5
Patch Set 5:

Seems there was a bug with how we get native threads in jvmti that I need to fix before we can merge this.

Basically if we were super unlucky we could end up with a thread being deleted as we were trying to notify on it. Since the deletion of the thread doesn't actually occur under the thread_list_lock_ we could end up hitting freed memory.

To solve this we need to wait for all threads to be done being destroyed.

Thanks for reviewing this one. I will rebase it on the fix to the other issue and submit it once that's in.

We were pretty lucky that TH caught this at all.

Patch-set: 5
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0