Experimental: ahat ART benchmark.
Implements ahat-benchmark.jar, suitable for use as an ART benchmark.

Moves java files specific to ahat.jar to jar-src.
Moves references to awt and javax Images to jar-src.

The benchmark loads ahat's test dump repeatedly for a minute and outputs
how much time it took on average to load the test dump. It takes around
6s to load the heap dump on host and 18s on bullhead.

Bug: 65236784
Test: m -j32 ahat-test
Test: art -cp ahat-benchmark.jar Main
Change-Id: I35a4bc6680bb415544a3646434ee509bb6e85bd0