Update patch set 2
Patch Set 2:

> I think we keep the sources of most of the benchmarks we
 > own/maintain under source control management. Shouldn't we do the
 > same with this one?

Where do we keep the sources for our benchmarks? If I submit this CL, then any changes to ahat in the future will change the benchmark as well. Would it make more sense to fork a copy of the current ahat code for the benchmark so the benchmark is stable while ahat continues to evolve?

Patch-set: 2
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>