Update patch set 2
Patch Set 2:

If it's pure .class java, you can check in the source code in our golem repository (or leave it here).

Otherwise you can check in the source code here, and a prebuilt in the golem repository (see go/dalvik -> Golem -> Adding new benchmark). 

The outer loop  (Main.java) should probably be checked into golem either way incase we need to change the benchmark harness logic.
-----
More often than not we don't update the benchmark code, in the rare case that we do it's usually an update to the harness.

Patch-set: 2