Update patch set 2
Patch Set 2: Code-Review+1

> ouch. how did this go unnoticed for so long?

Because users know that a failed close still closes. I'm not sure that not reporting an error is good. I mean, hey, that's what the kernel does itself, so.

And it also means that you now do *not* get a real error when you're trying to call Close() twice?

(Resetting the fd, on the other hand, is good sanitization)

Patch-set: 2
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1