Update patch set 1
Patch Set 1: Presubmit-Verified+1

=== Finished presubmit run: L42200000108589996 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/501540?revision=1




TreeHugger finished with: 12 passed.







WARNING: The tip-of-tree(4375952) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4375907), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







WARNING: The tip-of-tree(4375952) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4375840), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4375952) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4375922), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Bravo!

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1