Update patch set 2
Patch Set 2:

> I know at least half of us on the other side of the pond use the
 > 'art' script and I just thought it  might be surprising to everyone
 > if a lot of logs just disappeared?
 > 
 > Plus, is there a way to bring them back without editing the script
 > manually again?

What's in the logs that's so interesting to you?

Anyway, dropping the logs is one option. Removing the GC logs is another one, but those logs AFAIK are useful for actual devices.

If you want a better implementation, feel free to chime in :)

Patch-set: 2
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>