Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Maybe they don't *all need it*, but why can't we always run with the wrapper from runtest? Would it break anything?
> 
> Then there doesn't need to be 2 separate flags, it seems much simpler?

It would make it harder to test and use ones which don't need it, which is expected to become the default case eventually. Also a grand total of 12 lines are directly related to this flag (including the help message) so I really don't think it would make things much simpler.

Patch-set: 3
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>