Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)

Context: We are planning to use the output of dexoptanalyzer in frameworks to get the information about the dex opt status including compiler filter. For the primary ones, we already have exposed an API in dalvik.DexFile, however, it cannot be used for secondary dex files due to selinux restrictions. We wanted to use dexoptanalyzer to do that job for us.

Patch-set: 4
Reviewer: Gerrit User 1084268 <1084268@85c56323-6fa9-3386-8a01-6480fb634889>