Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> That sounds like an agent issue and could lead valgrind/ASAN to complain about leaks. I'd rather you fix libjdwp.

I would too but the issue seems to be with spurious wakeups of RawMonitorWait and it's not clear that it's really fixable in a good way. Also we cannot really do that for prebuilts and having *every* test as a known failure is not very helpful.

Patch-set: 1
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>