Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not sure if will scale?
> 
> Today we need -Xmx; what if tomorrow -Xms, the next week -Xss: are we just going to add a new flag every time?

This is a test fix for a minor issue on a specific device. There are strong reasons to think that adjusting the stack size will not be needed on any device as a system-wide adjustment (individual tests can be handled with the 'run' scripts).

I also really don't see exactly what is wrong with adding flags to control things that you often want to control. That is entirely what they are for. By that logic literally every flag in run-test could be replaced with some inhumanely long combination of --runtime-flag and -Xcompiler-option.

Also what scaling problems are you worried about. This entire thing is just a shell script for executing other (much longer runnning) things.

Patch-set: 1
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>