Update patch set 3
Patch Set 3:

(2 comments)

I think this is a substantial improvement in that functions that quietly release and acquire a lock are really dangerous, and we no longer have that. But I'm not too enthusiastic about functions that release a not-obviously-related lock. It already confused the reviewers here, and I think this will keep happening.

I'd like to see something that makes it clear in client code that RequestSynchronousCheckpoint releases a lock. Renaming it to RequestSynchronousCheckpointAndReleaseThreadListLock seems a bit over the top. Adding a comment at every call might work if all else fails. 

Maybe invent a version of MutexLock that tracks the locked state like unique_lock (as Andreas already suggested), and explicitly pass that in to RequestSynchronousCheckpoint, as you will presumably have to anyway? That at least makes the relationship clear.

Patch-set: 3
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>