Update patch set 1
Patch Set 1:

> Patch Set 1: Presubmit-Ready+1
> 
> > Patch Set 1:
> > 
> > Andreas, is there anyone else who might be interested in this change?
> > 
> > This will make a slight change in anr traces, whereas before, you saw one frame in the odex or oat file, now there are no frames in that. It looks like this now:
> > 
> >   native: #00 pc 000000000001de6c  /system/lib64/libc.so (syscall+28)
> >   native: #01 pc 00000000000e9500  /system/lib64/libart.so (art::ConditionVariable::WaitHoldingLocks(art::Thread*)+152)
> >   native: #02 pc 0000000000254934  /system/lib64/libart.so (art::gc::TaskProcessor::GetTask(art::Thread*)+368)
> >   native: #03 pc 00000000002551fc  /system/lib64/libart.so (art::gc::TaskProcessor::RunAllTasks(art::Thread*)+92)
> >   at dalvik.system.VMRuntime.runHeapTasks(Native method)
> >   at java.lang.Daemons$HeapTaskDaemon.runInternal(Daemons.java:461)
> >   at java.lang.Daemons$Daemon.run(Daemons.java:103)
> >   at java.lang.Thread.run(Thread.java:764)
> 
> What about mixed traces (JNI transitions)? Won't they now be a lot more cryptic?
> 
> What's the overall motivation for this change?

The problem is we added this to skip the compressed sections in generated code because they are huge. Unfortunately, we are using the compressed sections in some system libraries to contain other data (symbol tables for now), so I didn't want to use an option that disabled using compressed sections completely.

I could change the semantics of the way this works so that you get the first frame in these special maps. I do admit, that I've never liked seeing that frame, but I don't consume this data.

I don't think mixed traces work now, if they do, they should be the same. The new unwinder and the current libbacktrace interface should allow you to handle mixed traces, but I've never really had an easy way to test something like that. If there was some app that generated a mixed trace, I would be happy to make sure it works.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>