Fix DexInstructionIterator overrun bug
Changed the iterator to contain the dex PC and number of
instructions. When the iterator runs past the end of the code items,
it gets set to a special "end" iterator.

This fixes a bug where dex files with code items that ended in the
middle of instructions caused iterators to keep iterating and never
stop.

Bug: 67104794
Test: test-art-host

Change-Id: Idf7d51344659b2c75207bdf444e39f271feb8d3a