Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > So we're solidifying the broken annotation semantics here?
> 
> It has a Build visibility, so I do not see why we should resolve it at runtime.

Because that is how annotations work - they're relative to the class' classloader. So your commit message should at least mark that this means we are definitely deciding to violate regular semantics (rpl would know whether we had an open bug on that or not).

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>