Makes some ART gtests more robust to a non-standard environment.
Two gtests (dexdump_test and dexlist_test) used to try to write
a file in the root directory and expected this operation to fail.
However, in a non-standard environment (e.g. a testing context
where such a gtest is allowed to write in `/`), these tests would
not fail a expected. Replace these paths with more explicit ones
not depending on writing permissions.

Test: Run ART gtest dexdump_test and dexlist_test in a chroot on device.
Bug: 34729697
Bug: 68125496
Change-Id: I23537757639721133d694d5366cf05d9009f05ad