ART: Add StringBuilder.append() intrinsics.
And simplify their use knowing that they "return this".
This prepares for optimizing the String append pattern.

Note that intrinsics are never inlined but these methods are
defined in core-oj and therefore the rest of the boot image
and prebuilts are not allowed to inline them anyway.

Boot image sizes for aosp_taimen-userdebug:
 - before:
   arm/boot*.oat: 19756400
   arm64/boot*.oat: 23415656
 - after:
   arm/boot*.oat: 19653872 (-100KiB)
   arm64/boot*.oat: 23292784 (-120KiB)

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: aosp_taimen-userdebug boots
Test: run-gtests.sh
Test: testrunner.py --target --optimizing
Bug: 19575890
Change-Id: I2f3bac17830ac1c51287b090a96facddd3d09f5c