[WIP] ART: Optimize String append pattern.
TODO: Should we support StringBuilder.append(char[])? It
requires potentially throwing NPE at the wrong place. And
it's weird that AbstractStringBuilder.append(char[]) does
not append "null" for null input as the other overloads do.

Recognize String appending with StringBuilder and replace
the entire expression with a runtime call that perfoms the
append in a more efficient manner.

For now, require the entire pattern to be in a single block
and be very strict about the StringBuilder environment uses.
Also, do not accept Object/float/double arguments as they
would require a call from the entrypoint back to a helper
function in Java; this shall be implemented later.

Boot image size for aosp_taimen-userdebug:
 - before:
   arm/boot*.oat: 19653872
   arm64/boot*.oat: 23292784
   oat/arm64/services.odex: 22408664
 - after:
   arm/boot*.oat: 19432184 (-?KiB)
   arm64/boot*.oat: 22992488 (-?KiB)
   oat/arm64/services.odex: 22376776 (-?KiB)
Note that const-string in compiled boot image methods cannot
throw, but for apps it can and therefore its environment can
prevent the optimization for apps. We could implement either
a simple carve-out for const-string or generic environment
pruning to allow this pattern to be applied more often.

TODO: Measure performance impact

Bug: 19575890
Test: 697-checker-string-append
Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: aosp_taimen-userdebug boots.
Test: run-gtests.sh
Test: testrunner.py --target --optimizing
Change-Id: I51789bf299f5219f68ada4c077b6a1d3fe083964