[WIP] ART: Optimize String append pattern.
Recognize String appending with StringBuilder and replace
the entire expression with a runtime call that perfoms the
append in a more efficient manner.

For now, require the entire pattern to be in a single block
and be very strict about the StringBuilder environment uses.
Also, do not accept Object/float/double arguments as they
would require a call from the entrypoint back to a helper
function in Java; this shall be implemented later.

TODO: Update measurements. The following is includes
      impact of introducing the additional String.append()
      intrinsics.
Boot image size for aosp_taimen-userdebug:
 - before:
   arm/boot*.oat: 19900624
   arm64/boot*.oat: 23619616
   oat/arm64/services.odex: 22162616
 - after:
   arm/boot*.oat: 19506156
   arm64/boot*.oat: 23131496
   oat/arm64/services.odex: 22121224
Note that const-string in compiled boot image methods cannot
throw, but for apps it can and therefore its environment can
prevent the optimization for apps. We could implement either
a simple carve-out for const-string or generic environment
pruning to allow this pattern to be applied more often.

TODO: Measure performance impact

Test: 689-checker-string-append
Change-Id: I51789bf299f5219f68ada4c077b6a1d3fe083964