[WIP] Use intrinsic codegen for compiling intrinsic methods.
TODO: Native intrinsic methods.

When compiling an intrinsic method, create a fake code item
that invokes the same method and try to compile it. If the
call is actually intrinsified (or simplified to other HIR)
and yields a leaf method, use the result of this compilation
attempt, otherwise compile the actual code item.

We would ideally want to use the intrinsic codegen for all
intrinsics that do not generate a slow-path call to the
default implementation. Relying on the leaf method is
suboptimal as we're missing out on methods that do other
types of calls, for example runtime calls. This shall be
fixed in a subsequent CL.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 67717501
Change-Id: I640fda7c22d4ff494b5ff77ebec3b7f5f75af652