Update patch set 2
Patch Set 2:

Discussed with Dan offline.

Doing it in the build-side is quite complicated because the environment variables or source tree could change out under you, and this would have to invalidate the cache.

For ART it's actually a *feature* that the get_build_var would stay the same if something external would've normally triggered them to change.

In short we should add the build_var cache on the ART side (once in python, once in bash) and change all our scripts to use it.

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>