Add fd-forwarding transport lib
dt_fd_forward is a jdwpTransport that takes as an address a local
socket file-descriptor and uses it to allow some other piece of code
or program to control the actual attach and communication process.

tools/dt_fds_forward.py is a python program that can be used as a
controller for this comms system. This is useful for testing. It
implements the same behavior (excepting DDMS) that the adbconnection
plugin will but is capable of being easily used on normal host
machines. This works by passing down fds to its child processes.

The overall goal of this transport is to use it to allow us to safely
multiplex the out-bound data with DDMS data. This is needed to let us
match current DDMS behavior which transmits packets in the blind on
the same channel JDWP traffic is sent on.

Test: ./tools/dt_fds_forward.py  \
        ./test/run-test --host --dev 001-HelloWorld
Test: jdb -attach localhost:12345
Tset: telnet localhost 12345
Test: jdb -attach localhost:12345, detach, attach

Test: ./tools/dt_fds_forward.py  \
        --debug-lib $JAVA_HOME/jre/lib/amd64/libjdwp.so \
        -- \
        ./test/run-test --host --jvm --dev 001-HelloWorld
Test: jdb -attach localhost:12345

Bug: 62821960
Bug: 69169846

Change-Id: I654db6c6991c006933e1e1f0a4e41c13f795f9a8