Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is this simply re-running with new version of clang (and now it's adding spaces)?

Clang used to insert spaces into the macro expansion for the generator of this file. It no longer does this as of the latest toolchain version. So, yes, this seems to just be a re-run with the new clang toolchain. As far as I can tell this behavior changed in clang with https://reviews.llvm.org/D30427.

Patch-set: 2