MIPS32: Add compile time check for new entry points.
MIPS32 differentiates between functions which are called directly
vs. indirectly (see code comments for details). To correctly
generate code to call a function the boolean function
IsDirectEntrypoint() has to be updated to return the correct value
for the enum value which corresponds to the newly added function.
If a new case label for the new enum value isn't added to
IsDirectEntrypoint() this will cause a compile time error so that
the developer is alerted to the situation and can take corrective
action (again, see the code comments for details).

Previously, we had situations where new functions were added but
IsDirectEntrypoint() wasn't updated to return the correct boolean
value which caused hard to diagnose code failures. Hopefully,
these changes will prevent that class of bugs in the future.

Test: Boot on CI-20 hardware
Test: test/testrunner/testrunner.py --target --optimizing

Change-Id: Ief423e8f73c241a58d73a9527531549f04c19bbf