[DO NOT MERGE] Make LSE removing ConstructorFence only.
This CL makes following changes to LSE,
to help debug the failure of 667-checker-simd-alignment:

  - Allow LSE to analyze graphs that has SIMD.
  - Disable load elemination.
  - Disable store elemination.
  - Only allow LSE to remove ConstructorFence.

Is it safe to always remove ConstructorFences for NewArray which
never escape this thread?

A simple test0() function is added to 667-checker-simd-alignment to
help debug. NOTE: to reproduce/debug the issue '--gcstress' option is a must.

Test: ./art/test/run-test --optimizing --64 --gcstress 667-checker-simd-alignment

Change-Id: Ic300301f9845d400bc05e15a922803227b09860e