Update patch set 2
Patch Set 2:

So I doubt the constructor fence is the issue here.

I've been staring at the code the last hour, and had one question: we seem to assume (callee-saved) Q registers do not need to be saved across calls. However, the calling convention is such that the callee only saves the bottom 8 bytes. Is that an oversight in our vectorizer?

Patch-set: 2
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>