Update patch set 2
Patch Set 2:

We (Nicolas and me) believe this bug is due to vectorization + scheduling + treating Q8 as callee-save. Vectorizer creates a vector instruction (HVecDup?), scheduler reorders it with NewInstance, register allocator allocates Q8 (D8 is callee-save) for the vector instruction, NewInstance does not save the Q8 because D8 is callee save, the high bits of Q8 probably get cleared when the register is used as a narrower type somewhere in the NewInstance call (which is pretty heavy for gcstress) and we store zeroes to the array.

When we keep the ConstructorFence, the vector instruction remains after the NewInstance.

Patch-set: 2
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>