Update patch set 2
Patch Set 2:

CC (Concurrent Collector GC) should have strong enough invariants that removing a constructor fence shouldn't matter. 

Either the object gc-escapes into the thread after passing a suspend point (stronger than a ConstructorFence), or it is allocated after a suspend point (at which point it would already be in the To space and hence not scanned).

Unless we misunderstood the invariants or have a bug in their implementation :).

> We (Nicolas and me) believe this bug is due to vectorization + scheduling + treating Q8 as callee-save. 

Agreed, this seems more plausible especially if it only happens when SIMD is on.

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>