Update patch set 4
Patch Set 4:

So should I proceed with that or you'd prefer Mingyao to sort that? I have a similar WIP solution for typeconversion insertions in LSE.

BTW, should we still fix the type of ArraySets originated from fill-array-data in ReferenceTypePropogation? (the issue will not result in a bug as the LSE fix will tackle it)

Patch-set: 4
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>