Add support for alternate JDWP default implementations
Change jdwp options parsing to take place later and add a
-XjdwpProvider:_ option that can be used by the runtime or plugins to
select an appropriate jdwp implementation. The argument is a string.
If 'internal' is given the current internal jdwp implementation is
used. Other values must be handled by a plugin.

Also adds a runtime callback that will be invoked when the runtime
wants to start or stop the debugger (namely at the post-zygote fork
and just before exit).

Also add '-XjdwpOptions:_' in preparation for the eventual removal of
the existing -Xrunjdwp=_ and -Xagentlib:jdwp=_ options. All of these
options now store their arguments as a std::string to be interpreted
by the jdwp implementation as it sees fit.

Test: ./test.py --host -j50
Test: ./test/run-test --host --debug 001-HelloWorld
Bug: 62821960

Change-Id: Ie31db6b6f7d76a03d4ab8e178fcf298ed0eec203