Update patch set 5
Patch Set 5:

> > I believe vmarko meant do not create type conversion at the final
 > > generated code, not at IR level.
 > >
 > > Adding type conversion at each load isn't as efficient as adding
 > > type conversion at the store. We can make it work, but will be
 > more
 > > complicated. As I said, it only applies to hand crafted dex
 > > instructions. Code generated from source code havs the type
 > > conversion in place anyway. Well, instruction simplifier breaks
 > > that somewhat, but not by much.
 > 
 > I'd prefer lazily inserting the type conversion when we eliminate
 > the load.

But then adding a type conversion at each load isn't as efficient as adding a type conversion at the only store. It's not easy to track which store generate the value and go back to patch the store to have the type conversion since LSE merge store values.

 > 
 > We also need to account for loading a type different from the
 > underlying field/array-element type, see https://android-review.googlesource.com/510198
 > .

Is that a correctness issue? As long as the store instruction stores the component type, is should be fine, right? I know the load instruction type can change from the field type, but we'll use the load instruction's type when its value is used. It looks fine to me.

Patch-set: 5
Reviewer: Gerrit User 1043514 <1043514@85c56323-6fa9-3386-8a01-6480fb634889>