Update patch set 1
Patch Set 1:

I suspect I just need to change the pc+580 (addr 0x61c) to a Regex expression

Is running this on an Angler device the only way to verify it locally? The closest I have is a Bullhead and it is not failing.

Patch-set: 1
Reviewer: Gerrit User 1080224 <1080224@85c56323-6fa9-3386-8a01-6480fb634889>