Update patch set 5
Patch Set 5:

Ok, the load side necessity to add type conversion came as a surprise. Given that, I'd be happy to do a load side only change.  Even though store side type conversion overhead is very low (160 or so temporary type conversions added for compiling core.oat on host, with a small optimization. without it's 1100 or so). But given the world of types we are in, we don't need them (I'd say some implicit assumptions on types are tricky and we need to be careful). Thanks for all the discussions, I'll send out another CL.

Patch-set: 5
Reviewer: Gerrit User 1043514 <1043514@85c56323-6fa9-3386-8a01-6480fb634889>