Update patch set 9
Patch Set 9:

Is there anything I should do to ease the review of the patch? (I understand it is a huge one) It would be hard to split it into more than one as the changes are pretty monolithic. Would you like to look at and get the results of "loop peeling to eliminate invariant exits" first?

Patch-set: 9
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>