Update patch set 3
Patch Set 3:

I see how this achieves the goal whilst being the least intrusive, but perhaps it would be worth going into some extra trouble for the sake of avoiding passing `targetArch` into the parser and having it discard information? It's not a big deal - if you want to do it like this, I won't object.

The way I'd do it is to have `__extractLine` extract the set of architectures from the START line and effectively have each test case defined on a subset of architectures, or `None` for all architectures that the test driver can choose from. It will need to be fixed up in a handful of places but the parser will return all the information from the test file.

Maybe you've tried this and I'm missing something?

Patch-set: 3
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>